Refactor shared logic into parent class
